XW-2817 | run search tests on mercury and mobile wiki
